requirements.txt
gradio
yt-dlp
ffmpeg-python
transformers
sentencepiece
openai-whisper
